SQL: Some changes on the class and tester
int -> double for getscore
getScore test now compares properly
fixed getScore to do what it's suppose to do.

+++ include/util/sql_util.h
@@ -11,7 +11,7 @@ public:
         bool init ( const QString& dbname );
         int addTeam ( const QString& team_name, const QString& school );
         int addUser ( const UserData& ud );
-        int getScore ( const QString& user_name );
+        double getScore ( const QString& user_name );
         void setScore ( const QString& user_name, int score );
         bool authenticate ( const QString& user_name, const QString& password );
 public slots:
